resumeiq/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
sentence-transformers>=2.7.0
spacy>=3.7.0
pdfplumber>=0.11.0
scikit-learn>=1.4.0
motor>=3.4.0
pymongo>=4.7.0
python-dotenv>=1.0.0
numpy>=1.26.0
